util separation of concerns
