+++ app/.gitignore
@@ -18,3 +18,6 @@ instance/
 # Miscellaneous
 .DS_Store
 .idea/
+
+wes_notes.md
+
