+++ app/models/content_generator.py
@@ -1,4 +1,9 @@
-from utils import stage_content, save_progress, load_progress, generate_completion
+from utils.generator_utils import (
+    stage_content,
+    save_progress,
+    load_progress,
+    generate_completion,
+)
 
 
 class ContentGenerator:
@@ -23,18 +28,6 @@ class ContentGenerator:
             self._initialize()
         elif self.mode == "completion" and completion_params:
             self.generate_completion(**completion_params)
-
-    # If you just need a completion and don't want to generate markdown files
-    #         completion_params = {
-    #               "prompt": "Your prompt here",
-    #               "engine": "text-davinci-002",
-    #               "temp": 0.7,
-    #               "max_tokens": 1024,
-    #               "n": 1,
-    #               "stop": None,
-    #               "freq_pen": 0,
-    #               "pres_pen": 0,
-    #            }
 
     def _initialize(self):
         stage_content(
