+++ wes_notes.md
@@ -1,0 +1,11 @@
+    # If you just need a completion and don't want to generate markdown files
+    #         completion_params = {
+    #               "prompt": "Your prompt here",
+    #               "engine": "text-davinci-002",
+    #               "temp": 0.7,
+    #               "max_tokens": 1024,
+    #               "n": 1,
+    #               "stop": None,
+    #               "freq_pen": 0,
+    #               "pres_pen": 0,
+    #            }
